requirements.txt
flask
flask_cors
flask_sqlalchemy
requests
sqlalchemy
python-dotenv
psycopg2-binary
dnspython
pandas
xlsxwriter
openpyxl